adding venues api and resource links
Links are a proposed feature right now and won't need to be included until much later. They are there to allow the API to be navigated without external documentation using HATEOAS (https://docs.microsoft.com/en-us/azure/architecture/best-practices/api-design#use-hateoas-to-enable-navigation-to-related-resources). The venues endpoint is proposed but hopefully is okay as it follows the rest of the style. The changelog is formalised into a type because I wrote a small script that outputs Typescript types for this for the frontend and needed it to be marked as a schema.