Adding file bindings as a unique type
Rather than trying to tack this on to files which could get a bit complicated I'm adding these as their own unique messages. This is due to the multi-event nature of file bindings. May turn out to be the wrong idea but hopefully this works still. This will still be contained under hermes but under a different routing key for clarity