Adding docker setup files: haproxy, keycloak and better gateway
The haproxy configuration allows keycloak, the file server and the gateway all to be served from one container which is handy. Especially allowing the file server and the gateway to run on the same port with different urls. The composite gateway dockerfile includes the frontend with the gateway allowing it to be served completely as one application. The alternative to this is to serve it separately with a proxy which should be done in local development, and if doing that you can skip this dockerfile.