Removing unused or invalid files, explained in message
Vagrantfile = the file is very outdated and currently exposes the wrong ports. I'll try and come up with some solution for working on windows soon but for the time being I'd just recommend using wsl2. It's not a great compromise but I'll come up with something soon hopefully and add it to a contributing file

test.sh = The integration tests in their current form have been removed so this file is useless (its compose file is being removed at the same time). This will eventually come back but not right now

run.sh = The build script isn't needed with the new dockerfiles so you can just run the docker command on its own. When I get around to it I'm probably going to consolidate all of these commands into a single script (so you can do like node uems.js run/test/build) which will make things better

Requirements Capture.md = out of date, not being followed, everything is now on asana for requirements and stuff and we're using user stories rather than this kind of requirements system

docker-compose.yml = replaced with docker-compose-hub.yml. The naming change is a bit annoying but may be changed over eventually. The new file is adapted from my local one and should bring the project up in a nice way.

docker-compose.test.yml = removed as the tests will be rebuilt. Additionally this isn't the best way for this file to be structured as we should be able to use docker-compose merging to add an additional service without replacing the entire file

build.sh = builds are run through docker now so this isn't required. It's not hte most efficient system but it works better imo